requirements.txt
kalshi-python
pandas
numpy
ta
python-dotenv
rich
requests
